Create README.md
Client Query Management system is a real time query handling system. Clients can raise a query and the date time, status and the descriptions are stored in DB and displayed. Support team can view it answer the queries from the client. It is updated in the DB and that will be displayed as well.